Add module configuration script
Module configuration process is the same for every module, so this
script is better to be placed here in order to avoid duplication
within every module project and simplify upgrading the script. Module
projects still contain simple scripts (which would probably be updated
much more seldom) just to download the hostApp and run this script.